corrdinates now are in degrees

+++ FitDataService.API/DTOs/RecordData/CoordinateDto.cs
@@ -3,6 +3,6 @@ namespace FitDataService.API.DTOs.RecordData;
 public record CoordinateDto
 {
     public DateTime Timestamp { get; set; }
-    public int? PositionLat { get; set; }
-    public int? PositionLong { get; set; }
+    public double? PositionLat { get; set; }
+    public double? PositionLong { get; set; }
 }
